requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.13
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
audioread==3.0.0
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
black==19.10b0
boto3==1.26.64
botocore==1.29.64
braceexpand==0.1.7
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.2
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.19
editdistance==0.6.2
einops==0.6.0
exceptiongroup==1.1.0
executing==1.2.0
faiss-cpu==1.7.3
fastapi==0.89.1
fasttext==0.9.2
ffmpeg==1.4
ffmpy==0.3.0
filelock==3.9.0
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
g2p-en==2.1.0
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
googletrans==3.1.0a0
gradio==3.4.0
grpcio==1.51.1
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.12.0
hydra-core==1.2.0
hyperframe==5.2.0
idna==2.10
ijson==3.2.0.post0
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflect==6.0.2
iniconfig==2.0.0
ipadic==1.0.0
ipykernel==6.21.1
ipython==8.9.0
ipywidgets==8.0.4
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jiwer==2.5.1
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kaldi-python-io==1.2.2
kaldiio==2.17.2
kiwisolver==1.4.4
latexcodec==2.0.1
Levenshtein==0.20.2
librosa==0.9.2
lightning-utilities==0.6.0.post0
linkify-it-py==1.0.3
llvmlite==0.39.1
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mecab-python3==1.0.5
mpmath==1.2.1
multidict==6.0.4
nemo-toolkit==1.15.0
nest-asyncio==1.5.6
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.13.0
OpenCC==1.1.6
orjson==3.8.5
packaging==23.0
pandas==1.5.3
pangu==4.0.6.1
parameterized==0.8.1
paramiko==3.0.0
parso==0.8.3
pathspec==0.11.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip-api==0.0.30
pipreqs==0.4.11
plac==1.3.5
platformdirs==2.6.2
pluggy==1.0.0
pooch==1.6.0
portalocker==2.7.0
progress==1.6
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pybtex==0.24.0
pybtex-docutils==1.0.2
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
pydub==0.25.1
Pygments==2.14.0
PyNaCl==1.5.0
pynini==2.1.5
pyparsing==3.0.9
pypinyin==0.48.0
pypinyin-dict==0.5.0
PySocks==1.7.1
pytest==7.2.1
pytest-runner==6.0.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.8.6
pytz==2022.7.1
PyYAML==5.4.1
pyzmq==25.0.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sacrebleu==2.3.1
sacremoses==0.0.53
scikit-learn==1.2.1
scipy==1.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.14.0
setproctitle==1.3.2
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.3.2.post1
sox==1.4.1
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
starlette==0.22.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
termcolor==2.2.0
text-unidecode==1.3
textdistance==4.5.0
texterrors==0.4.4
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.21.2
typed-ast==1.5.4
typer==0.7.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
wandb==0.13.9
wcwidth==0.2.6
webdataset==0.1.62
websockets==10.4
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.5
wrapt==1.14.1
yarg==0.1.9
yarl==1.8.2
youtokentome==1.0.6
zipp==3.12.0
